pyproject.toml
[tool.poetry]
name = "bot-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "bot_test" }]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
